requirements.txt
asgiref==3.5.2
dj-config-url==0.1.1
Django==4.1
django-ninja==0.19.1
mysqlclient==2.1.1
pydantic==1.9.1
python-decouple==3.6
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
